Use long long instead of long
git-svn-id: http://svn.cern.ch/guest/lhcb/Lbcom/trunk@132451 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 8199b5dba37c80fbab7223c1004bffb42426b444

+++ Lbcom/Rich/RichAlgorithms/src/RichVaryGasPressureTemperatureAlg.h
@@ -80,13 +80,13 @@ namespace Rich
     RadiatorData m_nomTemperature;
 
     /// Event count
-    unsigned long int m_nEvt;
+    unsigned long long m_nEvt;
 
     /// Pressure variation amplitude (Pa)
     double m_presVaryAmp;
 
     /// Pressure variation period (# events)
-    unsigned long int m_presVaryPeriod;
+    unsigned long long m_presVaryPeriod;
 
     /// Tolerance on the pressure measurement values (Pa)
     double m_presTol;
